+++ src/CommonSource/Utilities/ArchiveSort.h
@@ -54,7 +54,7 @@ class CArchiveSort {
 	public:
 		/* ====================  LIFECYCLE     ======================================= */
 		/*  constructor */ 
-		CArchiveSort ( string inputFilename, string outputFilename, unsigned int *readCounter, pthread_mutex_t *readCounterMutex, unsigned int medianFragmentLength );
+		CArchiveSort ( string inputFilename, string outputFilename, uint64_t *readCounter, pthread_mutex_t *readCounterMutex, unsigned int medianFragmentLength );
 		//~CArchiveSort ();
 		void Sort();
 
@@ -74,7 +74,7 @@ class CArchiveSort {
 		vector< string > _tempSortedFiles;
 		
 		unsigned int _alignedReadCacheSize;
-		unsigned int *_readCounter;
+		uint64_t *_readCounter;
 		pthread_mutex_t *_readCounterMutex;
 		unsigned int _medianFragmentLength;
 		//unsigned int _extendedFragmentLength;
